refactor: 매물 등록 폴더 구조 재구성 및 공통 컴포넌트로 변경
