+++ src/components/listings/BadgeLists.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { CATEGORY_OPTIONS } from "@/constants/propertyCategory";
+import { CATEGORY_OPTIONS } from "@/constants/property-category";
 
 import { OptionItem } from "@/types/property";
