+++ src/components/listings/ExtraConditions.tsx
@@ -1,4 +1,4 @@
-import { CATEGORY_OPTIONS } from "@/constants/propertyCategory";
+import { CATEGORY_OPTIONS } from "@/constants/property-category";
 
 import { OptionItem } from "@/types/property";
