+++ src/components/listings/create/common/ListingCreateClient.tsx
@@ -7,14 +7,14 @@ import { useListingStore } from "@/stores/useListingStore";
 import { useFunnel } from "@/hooks/common/useFunnel";
 
 import Funnel from "@/components/common/Funnel/Funnel";
-import ContractTerms from "@/components/listings/create/ContractTermsLayout";
-import CreateConfirm from "@/components/listings/create/CreateConfirm";
 import CreateSuccess from "@/components/listings/create/CreateSuccess";
-import ListingDescription from "@/components/listings/create/ListingDescription";
-import MovingCondition from "@/components/listings/create/MovingConditionLayout";
 import AddressPhoto from "@/components/listings/create/addressPhoto/AddressPhoto";
+import ContractTerms from "@/components/listings/create/contractTerms/ContractTermsLayout";
+import CreateConfirm from "@/components/listings/create/createConfirm/CreateConfirm";
+import ListingDescription from "@/components/listings/create/listingDescription/ListingDescription";
 import ListingDetails from "@/components/listings/create/listingDetails/ListingDetailsLayout";
 import ListingType from "@/components/listings/create/listingType/ListingType";
+import MovingCondition from "@/components/listings/create/movingConditions/MovingConditionLayout";
 
 import { FUNNEL_FLOW } from "@/constants/funnel-steps";
