+++ src/components/listings/create/listingDetails/FurnitureContent.tsx
@@ -1,5 +1,5 @@
 import { furnitureIconMap } from "@/constants/furniture-lists";
-import { CATEGORY_OPTIONS } from "@/constants/propertyCategory";
+import { CATEGORY_OPTIONS } from "@/constants/property-category";
 
 interface FurnitureContentProps {
   value: number[];
