+++ src/components/listings/create/listingDetails/InternalDetailsContent.tsx
