+++ src/constants/property-category.ts
