+++ src/utils/listing/create/answerHelpers.ts
@@ -5,7 +5,7 @@ import {
   CAPACITY_SHARE_MAP,
 } from "@/constants/capacity-options";
 import { RENT_TYPE_MAP, SHARE_TYPE_MAP } from "@/constants/housing-options";
-import { CATEGORY_OPTIONS } from "@/constants/propertyCategory";
+import { CATEGORY_OPTIONS } from "@/constants/property-category";
 
 import { ListingDetailsOption } from "@/types/createPropertyAnswer";
 import {
